ai-engine/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
assemblyai
cohere
huggingface_hub
scikit-learn
pandas
numpy
scipy
httpx
python-multipart
loguru
python-jose[cryptography]
passlib[bcrypt]
cors
python-docx
PyPDF2
weaviate-client
